Made the whole thing even more config driven; by establishing a main config coming from the command line that contains the name of all the other configs and files.
Added configs and data for 2015, so that I could see what sort of stddev we
get when we try to predict 2019 results based on 2015 data.

Made the probability output format output numbers scaled to something around
1000, instead of inverses.

Added a bit of debug info.

Removed BS from the votes from the outputs, because their 2019 data is pretty
weak sauce.