A pretty huge overhaul of the whole context system for expression eval.
The main goal here was to allow custom ranges; which I needed for
committee_data/apply_thresholds.cfg (this is pretty complex stuff).
After the changes here, the transform library now supports stuff like
"for all columns from the one after district to the end, do a transform
where we check if the column value is greater than the column with the same
name except with an extra _cmp suffix".

Doing this in a reasonably generic fashion required the ability to evaluate
range boundaries as expressions. And that, in turn, required us to be able
to evaluate things like a column name to a column index, without actually
having a row in hand; which required changing the context system.

Right now, the context is basically a mapping from names to either other
names (which we follow recursively), or to numbers; which ideally are
indices in the row we're looking at (if there's a row). And then there
was the complexity of how to handle the contexts for aggregations.

Anyway, the whole thing now works; and I was able to write threshold
application. And validate that I can indeed correctly reconstruct results
for the 2015 election (took those; because obviously the results of that
election would've been different if thresholds weren't applied).